Split suspense into lite extension package
