+++ packages/agent-sdk/tsconfig.json
@@ -18,6 +18,7 @@
     "noPropertyAccessFromIndexSignature": true,
     "typeRoots": ["./node_modules/@types", "../lite/node_modules/@types"],
     "paths": {
+      "@pumped-fn/lite-extension-suspense": ["../lite-extension-suspense/src/index.ts"],
       "@pumped-fn/lite": ["../lite/src/index.ts"]
     }
   },
