+++ packages/lite-extension-suspense/src/index.ts
@@ -1,0 +1,165 @@
+import { tag, type Lite } from "@pumped-fn/lite"
+
+type MaybePromise<T> = T | Promise<T>
+
+export interface SuspenseStepCounter {
+  next: number
+}
+
+export interface SuspenseStepKey {
+  taskId: string
+  runId: string
+  step: number
+}
+
+export type SuspenseStepEntry =
+  | {
+      status: "pending"
+      key: SuspenseStepKey
+      targetName: string
+      input: unknown
+      kind?: string
+    }
+  | {
+      status: "resolved"
+      key: SuspenseStepKey
+      targetName: string
+      value: unknown
+    }
+  | {
+      status: "completed"
+      key: SuspenseStepKey
+      targetName: string
+      result: unknown
+    }
+
+export interface SuspenseEventLog {
+  get(key: SuspenseStepKey): Promise<SuspenseStepEntry | undefined>
+  putPending(entry: Extract<SuspenseStepEntry, { status: "pending" }>): Promise<void>
+  putCompleted(entry: Extract<SuspenseStepEntry, { status: "completed" }>): Promise<void>
+  resolve(key: SuspenseStepKey, value: unknown): Promise<void>
+}
+
+export interface SuspenseExecEvent {
+  key: SuspenseStepKey
+  target: Lite.ExecTarget
+  ctx: Lite.ExecutionContext
+  targetName: string
+  input: unknown
+}
+
+export interface SuspenseExtensionOptions {
+  log: SuspenseEventLog
+  name?: string
+  defaultTaskId?: string
+  defaultRunId?: string
+  shouldHandle?: (target: Lite.ExecTarget, ctx: Lite.ExecutionContext) => boolean
+  shouldSuspend?: (event: SuspenseExecEvent) => MaybePromise<boolean>
+  run?: (event: SuspenseExecEvent, next: () => Promise<unknown>) => Promise<unknown>
+  getKey?: (ctx: Lite.ExecutionContext, target: Lite.ExecTarget) => SuspenseStepKey
+  getTargetName?: (target: Lite.ExecTarget, ctx: Lite.ExecutionContext) => string
+  createPendingEntry?: (event: SuspenseExecEvent) => Extract<SuspenseStepEntry, { status: "pending" }>
+}
+
+export const suspense = tag<boolean>({ label: "suspense.step", default: false })
+export const suspend = tag<boolean>({ label: "suspense.suspend", default: false })
+export const taskId = tag<string>({ label: "suspense.taskId" })
+export const runId = tag<string>({ label: "suspense.runId" })
+export const stepCounter = tag<SuspenseStepCounter>({ label: "suspense.stepCounter" })
+
+export class SuspendSignal extends Error {
+  override readonly name = "SuspendSignal"
+
+  constructor(readonly entry: SuspenseStepEntry) {
+    super(`Execution suspended at ${formatSuspenseStepKey(entry.key)}`)
+  }
+}
+
+export { SuspendSignal as SuspenseSignal }
+
+export function formatSuspenseStepKey(key: SuspenseStepKey): string {
+  return `${key.taskId}:${key.runId}:${key.step}`
+}
+
+export function createSuspenseContext(
+  scope: Lite.Scope,
+  options: {
+    taskId: string
+    runId: string
+    markers?: Lite.Tagged<any>[]
+  }
+): Lite.ExecutionContext {
+  return scope.createContext({
+    tags: [
+      taskId(options.taskId),
+      runId(options.runId),
+      stepCounter({ next: 0 }),
+      ...(options.markers ?? []),
+    ],
+  })
+}
+
+export function createSuspenseExtension(options: SuspenseExtensionOptions): Lite.Extension {
+  return {
+    name: options.name ?? "suspense",
+    async wrapExec(next, target, ctx) {
+      if (!(options.shouldHandle ?? shouldHandleSuspenseTarget)(target, ctx)) return next()
+
+      const key = options.getKey ? options.getKey(ctx, target) : nextSuspenseKey(ctx, options)
+      const targetName = options.getTargetName ? options.getTargetName(target, ctx) : getTargetName(target, ctx)
+      const event = { key, target, ctx, targetName, input: ctx.input }
+      const existing = await options.log.get(key)
+      if (existing?.status === "completed") return existing.result
+      if (existing?.status === "resolved") return existing.value
+      if (existing?.status === "pending") throw new SuspendSignal(existing)
+
+      const shouldSuspend = await (options.shouldSuspend ?? shouldSuspendTarget)(event)
+      if (shouldSuspend) {
+        const pending = options.createPendingEntry
+          ? options.createPendingEntry(event)
+          : { status: "pending" as const, key, targetName, input: ctx.input }
+        await options.log.putPending(pending)
+        throw new SuspendSignal(pending)
+      }
+
+      const result = await (options.run ? options.run(event, next) : next())
+      await options.log.putCompleted({ status: "completed", key, targetName, result })
+      return result
+    },
+  }
+}
+
+function shouldHandleSuspenseTarget(target: Lite.ExecTarget): boolean {
+  return typeof target !== "function" && (
+    suspense.find(target) === true ||
+    suspend.find(target) === true
+  )
+}
+
+function shouldSuspendTarget(event: SuspenseExecEvent): boolean {
+  return hasMarker(event.target, suspend)
+}
+
+export function hasMarker<T>(target: Lite.ExecTarget, marker: Lite.Tag<T, boolean>): boolean {
+  if (typeof target === "function") return false
+  return marker.find(target) === true
+}
+
+function nextSuspenseKey(
+  ctx: Lite.ExecutionContext,
+  options: Pick<SuspenseExtensionOptions, "defaultTaskId" | "defaultRunId">
+): SuspenseStepKey {
+  const foundTaskId = ctx.data.seekTag(taskId) ?? options.defaultTaskId ?? "default-task"
+  const foundRunId = ctx.data.seekTag(runId) ?? options.defaultRunId ?? "default-run"
+  let counter = ctx.data.seekTag(stepCounter)
+  if (!counter) {
+    counter = { next: 0 }
+    ctx.data.setTag(stepCounter, counter)
+  }
+  return { taskId: foundTaskId, runId: foundRunId, step: counter.next++ }
+}
+
+function getTargetName(target: Lite.ExecTarget, ctx: Lite.ExecutionContext): string {
+  if (typeof target === "function") return ctx.name ?? target.name ?? "anonymous"
+  return ctx.name ?? target.name ?? "anonymous"
+}
