+++ packages/lite-extension-suspense/tsconfig.json
@@ -18,8 +18,6 @@
     "noPropertyAccessFromIndexSignature": true,
     "typeRoots": ["./node_modules/@types", "../lite/node_modules/@types"],
     "paths": {
-      "@pumped-fn/agent-sdk": ["../agent-sdk/src/index.ts"],
-      "@pumped-fn/lite-extension-suspense": ["../lite-extension-suspense/src/index.ts"],
       "@pumped-fn/lite": ["../lite/src/index.ts"]
     }
   },
